feat: ホームとログイン画面を追加
2画面を行き来出来るところまで実装

初期生成されるカウンター周りは削除
ただテストコードは後で参考にするかもしれないので残している
