+++ lib/pages/home_page.dart
@@ -1,0 +1,27 @@
+import 'package:flutter/material.dart';
+
+class HomePage extends StatelessWidget {
+  const HomePage({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: const Text('Home'),
+      ),
+      body: Center(
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: [
+            ElevatedButton(
+              child: const Text('Login'),
+              onPressed: () {
+                Navigator.of(context).pushNamed('/login');
+              },
+            )
+          ],
+        ),
+      ),
+    );
+  }
+}
